(refactor) restructure files
- remove api-folder and move its files into lib/documents and plugins (where they fit better)
- add more index.ts files
- add .test-suffix to all test files to make them more distinguishable
