--- packages/language-server/src/api/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces';
-export * from './Document';
-export * from './fragmentPositions';
