+++ packages/language-server/src/lib/documents/index.ts
@@ -1,0 +1,6 @@
+export * from './DocumentManager';
+export * from './ManagedDocument';
+export * from './Document';
+export * from './TextDocument';
+export * from './Fragment';
+export * from './utils';
