+++ packages/language-server/src/plugins/index.ts
@@ -1,0 +1,6 @@
+export * from './css/CSSPlugin';
+export * from './typescript/TypeScriptPlugin';
+export * from './svelte/SveltePlugin';
+export * from './html/HTMLPlugin';
+export * from './PluginHost';
+export * from './interfaces';
