+++ packages/language-server/test/lib/documents/DocumentManager.test.ts
@@ -1,8 +1,7 @@
 import sinon from 'sinon';
 import * as assert from 'assert';
 import { TextDocumentItem, Range } from 'vscode-languageserver-types';
-import { TextDocument } from '../../../src/lib/documents/TextDocument';
-import { DocumentManager } from '../../../src/lib/documents/DocumentManager';
+import { TextDocument, DocumentManager } from '../../../src/lib/documents';
 
 describe('Document Manager', () => {
     const textDocument: TextDocumentItem = {
